% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_functions.R
\name{check_api_params}
\alias{check_api_params}
\title{Validate API parameters}
\usage{
check_api_params(.api_params)
}
\arguments{
\item{.api_params}{Named list taken from \code{api_params} argument in
\code{\link{get_fema_data}} \code{\link{get_fema_data_all}}}
}
\description{
Validate API parameters
}
\details{
Not meant for interactive use. Helper function used within \code{\link{get_fema_data_all}} and
\code{\link{get_fema_data_all}}
}
\seealso{
\code{\link{get_fema_data}} \code{\link{get_fema_data_all}}

Other helpers: 
\code{\link{query_fema}()}
}
\concept{helpers}
