% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_functions.R
\name{get_fema_data_all}
\alias{get_fema_data_all}
\title{Get FEMA Open API All Data}
\usage{
get_fema_data_all(
  entity,
  api_params = list(),
  max_limit = NULL,
  wait = 1,
  base_url = "https://www.fema.gov/api/open"
)
}
\arguments{
\item{entity}{Required. Character. Name of entity to scrape. See function \code{\link{get_fema_entities}}
documentation for details.}

\item{api_params}{Optional. Named list. Extra arguments to pass to api.
Parameters are validated using the \code{\link{check_api_params}} function.}

\item{max_limit}{Optional. Numeric. Single number of maximum cases to return. If default (\code{Null}),
return all records.}

\item{wait}{Number of seconds to process the data.}

\item{base_url}{Optional. Character string representing website Url of openFEMA API. By default is \code{https://www.fema.gov/api/open}.}
}
\value{
A list of data and meta information see \code{examples} for more details.
}
\description{
Wrapper around \code{\link{get_fema_data}}, however, provides the ability to return more than
1000 records.
}
\examples{
\dontrun{
# get all data from Disaster Declarations  where fiscal year is greater than 1979
all_data <- get_fema_data_all("Disaster Declarations Summaries", 
            api_params = list(filter = "fyDeclared gt 1979"))
# convert all data to single data frame
all_data_df <- lapply(all_data, function(x) {
 x <- x[[2]]
})

all_data_df <- do.call(rbind, all_data_df)
}
}
\seealso{
Other get_data: 
\code{\link{get_fema_data}()},
\code{\link{get_fema_entities}()}
}
\concept{get_data}
