% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_functions.R
\name{get_fema_entities}
\alias{get_fema_entities}
\title{Get FEMA Open Entity Names}
\usage{
get_fema_entities(
  entity,
  entity_url = "https://www.fema.gov/about/openfema/data-sets",
  verbose = FALSE
)
}
\arguments{
\item{entity}{Required. Character. Name of entity to scrape. See function \code{\link{get_fema_entities}}
documentation for details.}

\item{entity_url}{Optional. Character. A url string of the website containing FEMA entity names.
The default value is \code{https://www.fema.gov/about/openfema/data-sets}}

\item{verbose}{Logical. If \code{FALSE} (default), will provide information about \code{entity}. If \code{TRUE}, will return
all possible entities.}
}
\value{
If \code{verbose = TRUE} will return a data.frame of all entity names.
}
\description{
A function to return a list of FEMA entities available to query.
Scrapes \url{https://www.fema.gov/about/openfema/data-sets} for most up-to-date entity names.
}
\examples{
\dontrun{
# Return list of FEMA entities to query
get_fema_entities(verbose = TRUE)
}
}
\seealso{
Other get_data: 
\code{\link{get_fema_data_all}()},
\code{\link{get_fema_data}()}
}
\concept{get_data}
