% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{map_fema_data}
\alias{map_fema_data}
\title{Map FEMA data to Stacked Bar}
\usage{
map_fema_data(df, x_var = "state", by = "incidentType")
}
\arguments{
\item{df}{Required. Data.frame created by \code{\link{get_fema_data}} or
\code{\link{get_fema_data_all}}}

\item{x_var}{Optional Character string. Column name in \code{df} to be placed on x axis. state by default.}

\item{by}{Optional Character string. Column name in \code{df} to be used as fill.
incidentType by default.}
}
\value{
A stacked \code{geom_bar} from \code{ggplot2}
}
\description{
Map FEMA data to Stacked Bar
}
\examples{
\dontrun{
## Assign entity to be used to get data
entity <- "DisasterDeclarationsSummaries"

## get data when fiscal year is greater than 1979
FEMAdf <- get_fema_data(entity, api_params = list(filter = 'fyDeclared gt "1979"'))

## plot data
map_fema_data(FEMAdf[[2]])
}

}
\seealso{
\code{\link[ggplot2]{geom_bar}}

\code{\link{get_fema_data}} \code{\link{get_fema_data_all}} for getting fema data
}
