% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_functions.R
\name{query_fema}
\alias{query_fema}
\title{Query FEMA helper}
\usage{
query_fema(api_url, .wait = 1)
}
\arguments{
\item{api_url}{Character. Website URL or FEMA's API}

\item{.wait}{Number of seconds to process the data. If default, will wait 1 second between api calls.}
}
\value{
API response from \url{https://www.fema.gov/api/open}
}
\description{
Query FEMA helper
}
\details{
Not meant for interactive use. Helper function used within \code{\link{get_fema_data_all}} and
\code{\link{get_fema_data_all}}
}
\seealso{
Other helpers: 
\code{\link{check_api_params}()}
}
\concept{helpers}
