% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_functions.R
\name{get_fema_data}
\alias{get_fema_data}
\title{Get FEMA Open API Data}
\usage{
get_fema_data(
  entity,
  api_params = list(),
  base_url = "https://www.fema.gov/api/open"
)
}
\arguments{
\item{entity}{Required. Character. Name of entity to scrape. See function \code{\link{get_fema_entities}}
documentation for details.}

\item{api_params}{Optional. Named list. Extra arguments to pass to api.
Parameters are validated using the \code{\link{check_api_params}} function.}

\item{base_url}{Optional. Character string representing website Url of openFEMA API. By default is \code{https://www.fema.gov/api/open}.}
}
\value{
A dataset taken from \url{https://www.fema.gov/api/open}
}
\description{
A function to return 1000 records from a FEMA data entity \href{https://www.fema.gov/api/open}{openFEMA api}
}
\examples{
\dontrun{
## Get first 1000 records from FEMA Web Disasters
get_fema_data("FEMA Web Disasters", api_params = list(skip = 1000))

## Get first 1000 records from FEMA Web Disasters after skipping first 1000 records
get_fema_data("FEMA Web Disasters", api_params = list(skip = 1000))

## Get first 1000 records ordered by state from Disaster Declarations Summaries after skipping first 1000 records
get_fema_data("Disaster Declarations Summaries",
  api_params = list(orderby = state, skip = 1000)
}
}
\seealso{
\url{https://www.fema.gov/about/openfema/api}

Other get_data: 
\code{\link{get_fema_data_all}()},
\code{\link{get_fema_entities}()}
}
\concept{get_data}
